Commit 02 — migrate remaining non-legacy `TestBase` usages
